Add files via upload
BIGMART SALES PREDICTION:-

PROBLEM STATEMENT
  To analyze and predict the sales of bigmart using different machine learning algorithms.
The aim is to build a predictive model and find out the sales of each product at a particular store.Using this model,We will try to understand the properties of products and stores which play a key role in increasing sales

The data scientists at BigMart have collected 2013 sales data for 1559 products across 10 stores in different cities. Also, certain attributes of each product and store have been defined. The aim is to build a predictive model and predict the sales of each product at a particular outlet.
Using this model, BigMart will try to understand the properties of products and outlets which play a key role in increasing sales

DATASET EXPLANATION:-
Item_Identifier: Unique product ID
Item_Weight: Weight of product
ItemFatContent: Whether the product is low fat or not
Item_Visibility: The % of total display area of all products in a store allocated to the particular product
Item_Type: The category to which the product belongs
Item_MRP: Maximum Retail Price (list price) of the product
Outlet_Identifier: Unique store ID
OutletEstablishmentYear: The year in which store was established
Outlet_Size: The size of the store in terms of ground area covered
OutletLocationType: The type of city in which the store is located
Outlet_Type: Whether the outlet is just a grocery store or some sort of supermarket
ItemOutletSales: Sales of the product in the particular store. This is the outcome variable to be predicted
 
Objectives of this project :-
Perform data cleaning on the dataset
Make a EDA report
Visualize the distributions of various features and correlations between them
Feature engineering to extract the correct features for the model
Build a regression model to predict the outlet sales

Let’s evaluate 5 different algorithms:
Logistic Regression (LR)
K-Nearest Neighbors (KNN).
DecisionTree (DECT).
Naive Bayes (NB).
Support Vector Machines (SVM).

In this project excersie, i learn to use a step-by-step on machine learning project using Python.
Based on the study, sales predict problem can be examined by using machine learning metholodgies across different alogrithms.